Create Chat bot to known about different  personalities in different streams of
Now a days we enjoys a lot many new inventions and with the ideology of many different people and their ways. Many people are anxiety to known that well personalies names and thier discoveries in thier different streams. My chat bot helps them to know the names and their discoveries in different streams.